Added a ton of weapon images. Inventory now rotates weapon down 45 degrees Grabbing an item now rotates it up through to 25 degrees + wobble Object cloning now correctly copies text data and materials Moved description builder from fpslogic to hv.equipment.info.english.cpp Moved StringBuilder << (char c) to StringBuilder.Append(char c)  ..which removed some bad casts

+++ hyperverse/hv.script.types.h
@@ -63,7 +63,7 @@ namespace HV {
 namespace HV { 
 	struct InventoryData
 	{
-		int32 legalEquipmentSlotFlags;
+		int64 legalEquipmentSlotFlags;
 		int32 maxStackSize;
 		ID_SPRITE icon;
 	};
